[hpiutil] Rename /sysvar.(hpp|cpp)/Sysvar.$1/
