+++ src/hpiutil/Sysvar.cpp
